+++ app/assets/scripts/editor.js
@@ -230,8 +230,10 @@
                         'rumble' + (c6Defines.kLocal ?
                             ('/app/index.html?kCollateralUrl=' +
                                 encodeURIComponent('../c6Content') +
-                                '&kDebug=true&kDevMode=true') :
-                            ('?kCollateralUrl=' + encodeURIComponent(c6Defines.kCollateralUrl) +
+                                '&kDebug=true&kDevMode=true' +
+                                '&kDevice=' + encodeURIComponent(this.device) +
+                                '&kMode=' + encodeURIComponent(experience.mode)) :
+                            ('/?kCollateralUrl=' + encodeURIComponent(c6Defines.kCollateralUrl) +
                                 '&kDevice=' + encodeURIComponent(this.device) +
                                 '&kMode=' + encodeURIComponent(experience.mode))
                         )
@@ -310,12 +312,11 @@
                 });
 
                 $scope.$on('mrPreview:updateMode', function(event, exp) {
-                    // the EditorCtrl $broadcasts when the experience
-                    // when the mode (full, light, etc) changes
+                    // the EditorCtrl $broadcasts the experience
+                    // when the mode (full, light, etc) changes.
                     // we need to convert and save the updated
-                    // experience and then tell the player to reload
+                    // experience, this will trigger a refresh automatically
                     experience = MiniReelService.convertForPlayer(exp);
-                    session.ping('mrPreview:updateMode');
                 });
 
                 $scope.$on('mrPreview:reset', function() {
@@ -326,15 +327,10 @@
                     return self.device;
                 }, function(newDevice, oldDevice) {
                     if(newDevice === oldDevice) { return; }
-
+                    // we longer have to tell the player that the mode changed
+                    // the iframe src will update and trigger a refresh automatically
+                    // we just prepare the profile for the refresh handshake call
                     profile.device = newDevice;
-
-                    // ping the MR player
-                    // sending 'updateMode' will trigger a refresh
-                    // and the player will call for another handshake
-                    // and will call for a specific card
-                    // in case we're previewing that card
-                    session.ping('mrPreview:updateMode');
                 });
             });
         }])