+++ CHANGELOG.md
@@ -10,3 +10,5 @@
 * [FIX]: Use min-safe DI in animation
 * [FIX]: Add kMode and kDevice query params to preview player
 * [FIX]: Remove "viewChangeStart" event handler when c6-view is destroyed, prevent memory leak
+* [FIX]: Add trailing slash for passing query params to preview player
+* [FIX]: Use preview player src to trigger refresh for mode/device changes